local snap-sync
