+++ packages/world/ts/plugins/snapsync/config.ts
@@ -1,0 +1,7 @@
+import { z } from "zod";
+
+export const zSnapSyncPluginConfig = z
+  .object({
+    snapSync: z.boolean(),
+  })
+  .catchall(z.any());
