+++ packages/world/ts/plugins/snapsync/index.ts
@@ -1,7 +1,3 @@
-/**
- * Importing this file automatically adds the required MUD modules to make
- * snap sync work.
- */
-
-import "./configExtensions";
-import "./typeExtensions";
+export * from "./config";
+export * from "./plugin";
+export * from "./types";
